Debugged system
The only debugging needed was to change mysql.connector.connect to mysql.connect on line 432, as mysql.connector was imported as mysql. The only other thing needed is to download the mysql software and then either use the username and password "root" or change the username and password in the program. 
The mysql software can be downloaded here: https://dev.mysql.com/downloads/installer/
A good tutorial if you are in need of guidance: https://www.youtube.com/watch?v=3vsC05rxZ8c