Integrated Program with GUI
Has the html system working with the backend system, just needs to be debugged and then the rest of the features can be added. You'll need the eel module to run it, and I think the file structure is important because it tells eel where to get the html files.